# Conflicts:
#	src/instr/instr_TI_trace.cpp
#	src/smpi/smpi_base.cpp
#	src/smpi/smpi_global.cpp
#	src/smpi/smpi_pmpi.cpp
#	src/smpi/smpi_replay.cpp

+++ teshsuite/smpi/mpich3-test/datatype/longdouble.c
@@ -35,8 +35,8 @@ int main(int argc, char *argv[])
         if (MPI_LONG_DOUBLE != MPI_DATATYPE_NULL) {
             MPI_Type_size(MPI_LONG_DOUBLE, &type_size);
             if (type_size != sizeof(long double)) {
-                printf("type_size != sizeof(long double) : (%zd != %zd)\n",
-                       (size_t)type_size, sizeof(long double));
+                printf("type_size != sizeof(long double) : (%d != %zd)\n",
+                       type_size, sizeof(long double));
                 ++errs;
             }
         }
@@ -45,8 +45,8 @@ int main(int argc, char *argv[])
         if (MPI_C_LONG_DOUBLE_COMPLEX != MPI_DATATYPE_NULL) {
             MPI_Type_size(MPI_C_LONG_DOUBLE_COMPLEX, &type_size);
             if (type_size != sizeof(long double _Complex)) {
-                printf("type_size != sizeof(long double _Complex) : (%zd != %zd)\n",
-                       (size_t)type_size, sizeof(long double _Complex));
+                printf("type_size != sizeof(long double _Complex) : (%d != %zd)\n",
+                       type_size, sizeof(long double _Complex));
                 ++errs;
             }
         }
@@ -62,4 +62,3 @@ int main(int argc, char *argv[])
     MPI_Finalize();
     return 0;
 }
-
